index.js now initializes everything
