+++ betterrolls5e/module.json
@@ -44,12 +44,7 @@
 		}
 	],
 	"systems": ["dnd5e", "dnd5eJP"],
-	"esmodules": [
-		"./scripts/betterrolls5e.js",
-		"./scripts/custom-roll.js",
-		"./scripts/init.js",
-		"./scripts/hooks.js"
-	],
+	"esmodules": ["./scripts/index.js"],
 	"styles": ["./css/betterrolls5e.css"],
 	"packs": [],
 	"url": "https://github.com/RedReign/FoundryVTT-BetterRolls5e/tree/master/betterrolls5e",
